+++ tests/integration/test_bill.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from quickbooks.objects.base import Ref
+from quickbooks.objects.base import Ref, Address
 from quickbooks.objects.bill import Bill
 from quickbooks.objects.detailline import AccountBasedExpenseLine, AccountBasedExpenseLineDetail
 from quickbooks.objects.vendor import Vendor
@@ -30,6 +30,15 @@ class BillTest(QuickbooksTestCase):
         vendor = Vendor.all(max_results=1, qb=self.qb_client)[0]
         bill.VendorRef = vendor.to_ref()
 
+        # Test undocumented VendorAddr field
+        bill.VendorAddr = Address()
+        bill.VendorAddr.Line1 = "123 Main"
+        bill.VendorAddr.Line2 = "Apartment 1"
+        bill.VendorAddr.City = "City"
+        bill.VendorAddr.Country = "U.S.A"
+        bill.VendorAddr.CountrySubDivisionCode = "CA"
+        bill.VendorAddr.PostalCode = "94030"
+
         bill.save(qb=self.qb_client)
 
         query_bill = Bill.get(bill.Id, qb=self.qb_client)
@@ -37,3 +46,13 @@ class BillTest(QuickbooksTestCase):
         self.assertEqual(query_bill.Id, bill.Id)
         self.assertEqual(len(query_bill.Line), 1)
         self.assertEqual(query_bill.Line[0].Amount, 200.0)
+
+        self.assertEqual(query_bill.VendorAddr.Line1, bill.VendorAddr.Line1)
+        self.assertEqual(query_bill.VendorAddr.Line2, bill.VendorAddr.Line2)
+        self.assertEqual(query_bill.VendorAddr.City, bill.VendorAddr.City)
+        self.assertEqual(query_bill.VendorAddr.Country, bill.VendorAddr.Country)
+        self.assertEqual(query_bill.VendorAddr.CountrySubDivisionCode, bill.VendorAddr.CountrySubDivisionCode)
+        self.assertEqual(query_bill.VendorAddr.PostalCode, bill.VendorAddr.PostalCode)
+
+
+