+++ tests/unit/objects/test_companycurrency.py
@@ -1,10 +1,8 @@
-from datetime import datetime
-
+import unittest
 from quickbooks.objects.companycurrency import CompanyCurrency
-from tests.integration.test_base import QuickbooksUnitTestCase
 
 
-class CompanyCurrencyTest(QuickbooksUnitTestCase):
+class CompanyCurrencyTest(unittest.TestCase):
     def test_unicode(self):
         company_currency = CompanyCurrency()
         company_currency.Name = "test"
@@ -22,3 +20,4 @@ class CompanyCurrencyTest(QuickbooksUnitTestCase):
         self.assertEqual(ref.name, "test")
         self.assertEqual(ref.type, "CompanyCurrency")
         self.assertEqual(ref.value, 23)
+