+++ CHANGELOG.rst
@@ -1,5 +1,15 @@
 Changelog
 =========
+* 0.9.6 (January 2, 2024)
+    * Replace RAuth with requests_oauthlib
+    * Removed python 2 code from client.py
+    * Removed unused dependencies from Pipfile
+    * Added new fields to Employee object
+    * Added VendorAddr to Bill object
+    * Added new fields to Estimate object
+    * Fix TaxInclusiveAmt and vendor setting 1099 creation
+    * Updated readme and contributing
+
 * 0.9.5 (November 1, 2023)
     * Added the ability to void all voidable QB types
     * Added to_ref to CreditMemo object